requirements.txt
PySide6  # Built into Python but listed here for completeness
pyserial
matplotlib
pytest